pyproject.toml
[project]
    name = "b2eb"
    version = "0.1.0"
    description = "tools/scripts to convert book to epub when other options have problems :)"
    readme = "README.md"
    authors = [{ name = "graham", email = "graham.annett@gmail.com" }]
    requires-python = ">=3.12"
    dependencies = ["pyautogui>=0.9.54", "pillow>=11.0.0", "numpy>=2.1.3"]

[tool.ruff]
    line-length = 120
